Updates to compile with 1.7 compiler
[git-p4: depot-paths = "//depot/openwatcom/": change = 31650]

+++ bld/setupgui/c/nlmver.c
@@ -55,7 +55,7 @@
 static  void    LenToASCIIZStr(char *tobuf, char *frombuf)
 {
         memcpy(tobuf,frombuf+1,*frombuf);
-        tobuf[*frombuf] = NULL;
+        tobuf[(int)*frombuf] = NULL;
 }
 
 int     ReturnNLMVersionInfoFromFile
